Revisioned.

+++ truevfs-ext/truevfs-ext-jmx/src/main/java/net/java/truevfs/ext/jmx/JmxSyncStatisticsView.java
@@ -85,7 +85,7 @@ extends StandardMBean implements JmxSyncStatisticsMXBean {
     }
 
     @Override
-    public int getSyncOperations() {
+    public long getSyncOperations() {
         return getSync().getSequenceNumber();
     }
